requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.13.2
azure-functions-durable==1.2.2
tqdm==4.64.1
PyGithub==1.58.0
slack_bolt==1.16.2
pandas==1.5.3
dataframe-image==0.1.5
lxml==4.9.2
Pillow==9.4.0
